RJQ: [predictor] re-implementment
