+++ forte/data/batchers.py
@@ -26,6 +26,7 @@ from forte.data.types import DataRequest
 from forte.data.data_utils_io import merge_batches, batch_instances
 from forte.data.ontology.top import Annotation
 from forte.data.ontology.core import Entry
+from forte.data.converter import Feature
 
 __all__ = [
     "ProcessingBatcher",
@@ -142,6 +143,134 @@ class ProcessingBatcher(Generic[PackType]):
     @abstractmethod
     def default_configs(cls) -> Dict[str, Any]:
         raise NotImplementedError
+
+
+class FixedSizeDataPackBatcherWithExtractor(ProcessingBatcher):
+    def __init__(self, cross_pack: bool = True):
+        super().__init__(cross_pack=cross_pack)
+        self.instance_pool: List[Annotation] = []
+        self.feature_pool: List[Dict[str, Feature]] = []
+        self.pool_size = 0
+        self.batch_is_full = False
+
+    def initialize(self, config):
+        super().initialize(config)
+        self.scope = config["scope"]
+        self.feature_scheme = config["feature_scheme"]
+        self.batch_size = config["batch_size"]
+        self.instance_pool.clear()
+        self.feature_pool.clear()
+        self.pool_size = 0
+        self.batch_is_full = False
+
+    def convert(self, features_collection):
+        collections = {}
+        for features in features_collection:
+            for tag, feat in features.items():
+                if tag not in collections:
+                    collections[tag] = []
+                collections[tag].append(feat)
+
+        converted = {}
+        for tag, features in collections.items():
+            converter = self.feature_scheme[tag]["converter"]
+            data, masks = converter.convert(features)
+            converted[tag] = {
+                "data": data,
+                "masks": masks,
+                "features": features
+            }
+        return converted
+
+    def _should_yield(self) -> bool:
+        return self.batch_is_full
+
+    def flush(self) -> Iterator[Tuple[List[PackType],
+                    Optional[List[Annotation]], Dict]]:
+        r"""Flush the remaining data."""
+        if self.pool_size > 0:
+            yield (self.data_pack_pool, self.instance_pool,
+                    self.convert(self.feature_pool))
+            self.feature_pool = []
+            self.data_pack_pool = []
+            self.instance_pool = []
+            self.pool_size = 0
+
+    def get_batch(
+            self, input_pack: PackType, context_type: Type[Annotation],
+            requests: DataRequest) -> Iterator[Tuple[List[PackType],
+                    Optional[List[Annotation]], Dict]]:
+        r"""Returns an iterator of data batches."""
+        # cache the new pack and generate batches
+
+        for (batch, num) in self._get_data_batch(input_pack, context_type):
+            packs, instance, features = batch
+            self.data_pack_pool.extend(packs)
+            self.instance_pool.extend(instance)
+            self.feature_pool.extend(features)
+            self.pool_size += num
+
+            # Yield a batch on two conditions.
+            # 1. If we do not want to have batches from different pack, we
+            # should yield since this pack is exhausted.
+            # 2. We should also yield when the batcher condition is met:
+            # i.e. ``_should_yield()`` is True.
+            if not self.cross_pack or self._should_yield():
+                yield from self.flush()
+
+    def _get_data_batch(
+            self, data_pack: PackType, context_type: Type[Annotation],
+            requests: Optional[DataRequest] = None, offset: int = 0) \
+            -> Iterable[Tuple[Dict, int]]:
+        r"""Get data batches based on the requests.
+
+        Args:
+            data_pack: The data pack to retrieve data from.
+            context_type: The context type of the data pack.
+            requests: The request detail.
+            offset: The offset for get_data.
+
+        Returns:
+
+        """
+        packs: List[DataPack] = []
+        instances: List[Annotation] = []
+        features_collection: List[Dict[str, Feature]] = []
+        current_size = self.pool_size
+
+        for instance in data_pack.get(self.scope):
+            features = {}
+            for tag, scheme in self.feature_scheme.items():
+                features[tag] = scheme['extractor'].extract(data_pack, instance)
+            packs.append(data_pack)
+            instances.append(instance)
+            features_collection.append(features)
+
+            if len(instances) == self.batch_size - current_size:
+                self.batch_is_full = True
+                batch = (packs, instances,
+                        self.convert(features_collection))
+                yield (batch, len(instances))
+                self.batch_is_full = False
+                packs = []
+                instances = []
+                features_collection = []
+                current_size = self.pool_size
+
+        # Flush the remaining data.
+        if len(instances) > 0:
+            batch = (packs, instances, features_collection)
+            yield (batch, len(instances))
+
+    @classmethod
+    def default_configs(cls) -> Dict[str, Any]:
+        config = super().default_configs()
+        config.update({
+            "scope": None,
+            "feature_scheme": {},
+            "batch_size": 10
+        })
+        return config
 
 
 class FixedSizeDataPackBatcher(ProcessingBatcher[DataPack]):
